+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 * Filter form: align selects and submit button
 * Hide customize link
+* Image are now called dynamically with the name "WeRecruit Logo %s" where %s is the company name of the job offer in WeRecruit, instead of getting the image in the job offer
 
 ### 1.0.1: July 22nd, 2020
 * Remove Clndr dependency